Move Texture and TextureStore to bloom::graphics namespace
