+++ BloomFramework/include/Graphics/Texture.h
@@ -3,7 +3,7 @@
 #include "stdIncludes.h"
 #include <optional>
 
-namespace bloom {
+namespace bloom::graphics {
 	class BLOOMFRAMEWORK_API Texture {
 	public:
 		Texture() = default;
